try making it closable
